+++ issuer-service/src/main/java/ch/admin/bj/swiyu/issuer/service/management/CredentialManagementService.java
@@ -87,7 +87,7 @@ public class CredentialManagementService {
      * of the offer and maps the result to a DTO suitable for clients.</p>
      *
      * @param managementId the id of the management object
-     * @param offerId the id of the offer object
+     * @param offerId      the id of the offer object
      * @return a {@link CredentialInfoResponseDto} containing credential offer information and a deeplink
      * @throws ResourceNotFoundException if no credential with the given id exists
      */
@@ -194,7 +194,7 @@ public class CredentialManagementService {
      * the offer has expired.</p>
      *
      * @param credentialManagementId the id of the credential offer
-     * @param offerId the id of the offer
+     * @param offerId                the id of the offer
      * @return the {@link StatusResponseDto} representing the credential's current status
      * @throws ResourceNotFoundException if no credential with the given id exists
      */
@@ -420,6 +420,7 @@ public class CredentialManagementService {
                         .credentialManagementStatus(CredentialStatusManagementType.INIT)
                         .renewalResponseCnt(0)
                         .renewalRequestCnt(0)
+                        .metadataTenantId(applicationProperties.isSignedMetadataEnabled() ? UUID.randomUUID() : null)
                         .build());
 
         CredentialOffer entity = persistenceService.saveCredentialOffer(
@@ -435,7 +436,6 @@ public class CredentialManagementService {
                         .credentialValidUntil(requestDto.getCredentialValidUntil())
                         .credentialMetadata(toCredentialOfferMetadataDto(requestDto.getCredentialMetadata()))
                         .configurationOverride(toConfigurationOverride(requestDto.getConfigurationOverride()))
-                        .metadataTenantId(applicationProperties.isSignedMetadataEnabled() ? UUID.randomUUID() : null)
                         .credentialManagement(credentialManagement)
                         .build());