+++ CHANGELOG.md
@@ -12,8 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - New endpoint `/actuator/env` to retrieve configuration details.
-- New endpoint `/management/api/credentials/{credentialManagementId}/offers/{offerId}` to retrieve offer specific information.
-- New endpoint `/management/api/credentials/{credentialManagementId}/offers/{offerId}/status` to retrieve the status of the offer.
+- New endpoint `/management/api/credentials/{credentialManagementId}/offers/{offerId}` to retrieve offer specific
+  information.
+- New endpoint `/management/api/credentials/{credentialManagementId}/offers/{offerId}/status` to retrieve the status of
+  the offer.
 - Send callback on every credential offer status change.
 - Send callback on every credential management status change.
 - Added field `event_trigger` to callback request
@@ -27,7 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed weak unlinkability by rounding down the timestamps within issued credetials.
   Affected fields are iat, epx, and nbf.
 - Removed credential request errors ISSUANCE_PENDING to be aligned with the spec.
-- Deferred credential response when credential data is not ready is now 202 ACCEPTED
+- Fixed signed metadata using always the first key used, even when keys were rotated by issuers during renewals.- Deferred credential response when credential data is not ready is now 202 ACCEPTED
 - Deferred credential transaction_id will not change anymore during deferred flow
 
 ## 2.3.1